backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.23
pandas==2.1.4
gunicorn==21.2.0
PyJWT==2.8.0
